application/backend/api/requirements.txt
fastapi == 0.104.1
uvicorn[standard] == 0.24.0
requests == 2.31.0
pyjwt[crypto] == 2.8.0
Jinja2 == 3.1.2
boto3 == 1.29.2
pycryptodome == 3.19.0
pyyaml == 6.0.1
